gcc fixes
Former-commit-id: 2e20b71d9359d8303a563a23bea12d744dc61f5b

+++ src/thing_move.cpp
@@ -247,7 +247,7 @@ bool Thing::move(point future_pos, uint8_t up, uint8_t down, uint8_t left, uint8
     if (! must_attack) {
       if (! is_undead() && ! is_ethereal()) {
         if (up || down || left || right) {
-          if (pcg_random_range(0, 20) > get_stat_constitution()) {
+          if ((int) pcg_random_range(0, 20) > get_stat_constitution()) {
             if (pcg_random_range(0, 100) < 10) {
               decr_stamina();
             }
